refactor: 결과 출력 부분을 별도의 파일로 분리
